Initial commit
This consists of stolen printf and strtod implementations. strtod() is
from musl, so it's probably fairly accurate. The printf was apparently
made for embedded use; the non-float code seems to be fairly complete,
but the float code is redundant/missing. The plan is to replace the
float formatting code with musl's. musl's printf was not considered,
because it's excessively cryptic and not something I want to maintain.

In both cases, the code is fairly unmodified. I removed the printfs
other than snprintf() and vsnprintf(), libc strlen is used, and the test
suite does not use a gross fucked up 13kloc C++ WTF POS. The strtod core
is shared with scanf in musl, which results in some code artifacts.