printf_test: compactify
Change it so that there's 1 line per test instead of 3 in the common
case. Also a good opportunity to change the indentation to what I
prefer.

I haven't done this in the initial commit, because I wanted to have a
chance to run a diff against upstream. This will not be needed anymore
since future commits will not do massive changes on this file.