printf: use a buffer struct instead of passing them as parameters
Before this commit, every function in the printf core which wants to
output characters took various parameters for the output buffer state.
The return value was also involved with this. Change all this to a
simple struct.

This also replaces the output callback. (I'm not even sure how it was OK
to call a callback per character, that'll definitely be slower than
other approaches...) It does not mean the printf core can't be used e.g.
for callback output. Instead, a "flush" callback could be added (so the
buffer can be smaller than the full size of the expected output string).

This also makes the code more similar to musl, which uses its own FILE
struct as buffer, using the same technique as suggested above to
implement fprintf().