printf_test: add some tests from libc-testsuite
In particular, this adds tests for float formatting.

Apparently exposes some problems, which will be fixed later.

Oddly, some tests in it were incorrect (explicitly fixed in musl, but
not in the testsuite). At least some of these were due to later
clarifications in the C standard. This probably means the testsuite is
unmaintained. But the working tests are nice to have anyway.