printf_test: add test for int/long extreme values and some PRI macros
Test common limits. Some of these are redundant, but it's good to have
them in a more explicit way. Didn't test the test for when long is 32
bit.

Also test that some of the most used PRI macros work correctly. It's
unknown what happens on Windows, where there isn't a "C native" int 64
type, and [u]int64_t aliases to a compiler-specific type. But this test
should catch it if it's broken.