printf: redo integer formatting path
This makes it pass some of the added tests (not all of them failed
before). This was mostly about weird corner cases with '#'. (I guess '#'
was a bad addition to the standard. There are more things weird about it
than not.)

In my opinion, this makes the code more readable (but maybe that's only
because I rewrote it to match my internal brain logic better).

This does not convert numbers backward and reverse them on output
anymore. Instead we simply write it backwards to the buffer. This
preserves the property of the upstream code that it starts outputting
the least significant digits first (so small numbers require fewer
iterations), for whatever that is worth.

This calculates all padding lengths etc. in advance before output.
Effectively, this means the padding can be output with separate calls,
and the maximum possible padding size is not restricted to
PRINTF_NTOA_BUFFER_SIZE anymore. This is good, because the standard
apparently doesn't allow such limits. The buffer size is now set to the
worst case that is possible at all with numbers (which is guaranteed to
be enough, but an assert() checks this in case someone breaks these
assumptions by messing up the code).

This still leaves the duplicated _ntoa_long and _ntoa_long_long. Maybe
it's worth the trouble to avoid 64 bit arithmetic in the 32 bit case.
But the commit tries to reduce the amount of duplicated code.