printf: report format and overflow errors
If the written size exceeds INT_MAX, report an error. (This was done
before, but if the total output exceeded size_t, random things could
happen.) If there are unknown format specifiers, report an error. To
make sure these cases are caught, we write the broken format to the
output, and append the string "<error>".

The only thing I'm aware of that we don't check is parsing inline
precission/width digits. This will overflow if there are too many
digits. I'm too lazy to correct this (it's pretty obscure and only an
issue if you maybe dynamically generate format strings?), but still
avoid returning negative numbers to avoid excessive UB.