printf: add support for explicitly sized integer types
Standard C provides implementation defined explicitly sized types in
stdint.h, like int32_t. A user can't know what these types map to. For
example, int32_t could map to int or long, and int64_t sometimes even
maps to a non-standard type. This means the printf() user can't know
which format specifier needs to be used.

The standard C solution is to use the format string macros in
inttypes.h, such as PRId64 for int64_t. This is a pretty ugly and
inconvenient solution.

This commit adds an alternative solution by adding explicit length
specifiers to the format string syntax. It follows what Microsoft did.
(Amazingly. This may be the only thing Microsoft ever did right.) The
requirements on this was that the bit or byte size is explicitly in the
format string. Since Number digits are already used for field width and
precision, there needs to be a prefix character (one that doesn't
collide with existing length/conversion specifiers). Microsoft seemed to
have the same thought, and all requirements match.

This syntax can be easily extended to other types. For example, if
anyone ever invents a int640_t type, it could be specified as "I640",
and would not collide with "I64". This is because it matches the longest
type, and the conversion specifier following the length specifier cannot
be a digit.

The implementation directly fetches these as int32_t/int64_t, so we
don't have to make any assumptions about which type this maps to. Unlike
the Microsoft extension, we also support [u]int8_t and [u]int16_t for
completeness. These are fetches as int, as types smaller than int (which
is guaranteed to be at least 16 bit) are always promoted to int.

I didn't check whether this really matches Microsoft behavior.

Note that all this is worthless if the compiler doesn't support this
printf specifiers. And most likely gcc/clang never will. I'm just trying
to make a point, or something.