printf: I guess it's compatible to msys64
It seems the inttypes.h macros mostly work, both with and without
-D__USE_MINGW_ANSI_STDIO=1. If the define is given, PRId64 is "lld", and
snprintf() is redirected to their own broken code. Without it, PRId64 is
"I64d", and native MSVCRT printf is used. Both happen to work.

(I'm not sure if there's any technical UB, since I64d uses a
Microsoft-specific __int64 type that is probably different from long
long. But msys itself does the same thing. Without the define mentioned
above, PRId64 is I64d, but int64_t is still typedefed to long long.)

This was tested only in 64 bit mode.

Note that the printf tests don't pass with system printf in either mode
on msys64. MSVCRT is ancient anyway (because msys insists on using an
ancient DLL that Microsoft is not updating anymore). msys's own code
(which I heard is derived from newlib) seems to be somewhat
non-compliant.