printf: implement %r extension
See: http://austingroupbugs.net/view.php?id=800

That va_list can be a pointer brings some severe pain. The proposal
above also suggests the addition of a special va_arg macro that works
with va_list, but we can't emulate it without assuming system specific
things. To solve this, require the caller to use LIN_VA_LIST when
passing va_list arguments.

I don't quite understand why, but initializing a struct with va_list
does not work on Linux/amd64. The va_list pointer value changes. So
codemonkey it and try&error things until something happens to work
(which was changing it to a pointer). Sorry about that.